requirements.txt
scikit-learn 
sentence-transformers 
faiss-cpu 
openai 
langchain 
streamlit
pandas
numpy
matplotlib
